RTC Smart-Clock  M5StackC Stick
This is a basic RTC clock example this will display both 24 hour time as the main clock and long press of A button will trigger 12 hour time. Also from the main 24 hour clock B button can be used to display ENVII hat stats temps in C/F and so on. It will loop 4 times with trigger flag. Note: no event timers are used with this example.  When in 12 hour time long press B button as A sleep function. it also has A Random color wheel for text effects. this example will not be exploiting fast sprite loading.